GameObject references (goref)
Object references in the PDBL are now loaded as direct pointers to objects instead of ID numbers.

Thus, finally, when saved, game objects IDs in the PDBL are now corrected when an object is inserted, moved or removed in the object tree (because the object ID in the SPK depends on the position in the object tree), and the game no longer crashes in these cases. But there still some issues...

It is possible to change an object reference in the DBL by dragging an object from the object tree and dropping on an object reference (list) in the DBL section of the selected object window.