Add PSCR removal tool
I attempted to remove the PSCR automatically in previous removed commits, however it seems some scripts are missing as sdl files, and removing the PSCR does have some side effects.
For example, in C4_1, the intro cutscene has the sdl file missing, and without the PSCR, 47 does not say his line during the intro cutscene.
So for now, you have to remove the PSCR by clicking the item in the Tool menu, giving you a choice.